refactor(earn): move type imports to earn types
