+++ apps/kyberswap-interface/src/pages/Earns/hooks/useSupportedDexesAndChains.ts
@@ -1,5 +1,6 @@
 import { useMemo } from 'react'
-import { PoolQueryParams, useSupportedProtocolsQuery } from 'services/earn'
+import { useSupportedProtocolsQuery } from 'services/earn'
+import type { PoolQueryParams } from 'services/earn/types'
 
 import { MenuOption } from 'components/DropdownMenu'
 import useChainsConfig from 'hooks/useChainsConfig'
